backend/requirements.txt
# Core Django dependencies
Django>=4.2.0,<4.3.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Statistical computation - Essential packages only
numpy>=1.24.3
pandas>=2.0.1
scipy>=1.10.1
statsmodels>=0.14.0
scikit-learn>=1.2.2
matplotlib>=3.7.1
seaborn>=0.12.2

# Data validation and processing
pydantic>=2.0.0
openpyxl>=3.1.2
xlrd>=2.0.1
pyyaml>=6.0
python-dotenv>=1.0.0

# File handling
Pillow>=9.5.0

# Development and testing
pytest>=7.3.1
pytest-django>=4.5.2